Reorganize code, allow storing data to  multiple storage engines

+++ cryptostore/data/parquet.py
@@ -9,8 +9,8 @@ import os
 import pyarrow as pa
 import pyarrow.parquet as pq
 
-from cryptostore.aggregator.store import Store
-from cryptostore.aggregator.gc import google_cloud_write
+from cryptostore.data.store import Store
+from cryptostore.data.gc import google_cloud_write
 
 
 class Parquet(Store):
@@ -24,17 +24,18 @@ class Parquet(Store):
                 self.bucket = config['GCS']['bucket']
                 self.prefix = config['GCS']['prefix']
                 self.auth = config['GCS']['service_account']
-    
+                self.del_file = False if 'del_file' in config['GCS'] and config['GCS']['del_file'] == False else True
+
     def aggregate(self, data):
         names = list(data[0].keys())
         cols = {name : [] for name in names}
-        
+
         for entry in data:
             for key in entry:
                 cols[key].append(entry[key])
         arrays = [pa.array(cols[col]) for col in cols]
         table = pa.Table.from_arrays(arrays, names=names)
-        
+
         if self.data is None:
             self.data = table
         else:
@@ -50,4 +51,5 @@ class Parquet(Store):
             if self.prefix:
                 path = f"{self.prefix}/{path}"
             self._write(self.bucket, path, file_name, creds=self.auth)
-            os.remove(file_name)
+            if self.del_file:
+                os.remove(file_name)
